changes_3 03.10.23
