+++ tests/api_tests_pet_store/get.spec.ts
@@ -1,14 +1,12 @@
-import {test, expect} from '@playwright/test'
+import {test, expect} from '@playwright/test';
 import axios from 'axios';
-import { baseUrl } from '../Global_settings/constant';
-import {faker, simpleFaker} from '@faker-js/faker';
+import {faker, simpleFaker} from '@faker-js/faker'
+import { baseUrl } from '../global_settings/constant';
 
 
 let randomFirstName = faker.person.firstName();
 let randomSecondName = faker.person.lastName();
 let randomId = simpleFaker.string.numeric(10);
-let updatedFirstName = faker.person.firstName();
-let updatedSecondName = faker.person.firstName();
 let petId: number;
 
 
@@ -16,7 +14,7 @@ let postData = {
   "id": randomId,
   "category": {
     "id": 0,
-    "name": randomFirstName
+    "name": randomFirstName,
   },
   "name": randomSecondName,
   "photoUrls": [
@@ -34,40 +32,14 @@ let postData = {
 
 test.beforeAll(async () => {
     const response = await axios.post(baseUrl, postData);
-    expect(response.status).toBe(200);
     petId = response.data.id;
    });
 
 test.afterAll(async () => {
     await axios.delete(`${baseUrl}${petId}`);
-});
+  });
 
-
-test('PutRequest', async () => {
-    const response = await axios.put(baseUrl, 
-    { 
-      data: {
-        "id": petId,
-        "category": {
-          "id": 0,
-          "name": updatedFirstName
-        },
-        "name": updatedSecondName,
-        "photoUrls": [
-          "string"
-        ],
-        "tags": [
-          {
-            "id": 0,
-            "name": "string"
-          }
-        ],
-        "status": "available"}
-    });
-    expect(response.status).toBe(200);
-   });
-
-   test ('Get', async () => {
+test('Get', async () => {
     const response = await axios.get(`${baseUrl}${petId}`);
     expect(response.data.category.name).toEqual(randomFirstName);
     expect(response.data.name).toEqual(randomSecondName);
