+++ tests/api_tests_pet_store/post.spec.ts
@@ -1,17 +1,17 @@
 import {test, expect} from '@playwright/test'
 import axios from 'axios';
 import {faker, simpleFaker} from '@faker-js/faker';
-import { baseUrl } from '../Global_settings/constant';
+import { baseUrl } from '../global_settings/constant';
 
 
 let randomFirstName = faker.person.firstName();
 let randomSecondName = faker.person.lastName();
-let randomId = simpleFaker.string.numeric(10);
+let randomIdNumber = faker.number.int(10)
 let petId: number;
 
 
 let postData = {
-  "id": randomId,
+  "id": randomIdNumber,
   "category": {
     "id": 0,
     "name": randomFirstName
@@ -30,13 +30,17 @@ let postData = {
  };
 
 
-test.beforeAll(async () => {
+test.afterAll(async () => {
+  await axios.delete(`${baseUrl}${petId}`);
+});
+
+test('Post', async () => {
     const response = await axios.post(baseUrl, postData);
     petId = response.data.id;
+    expect(response.status).toBe(200);
+    expect(response.data.id).toEqual(randomIdNumber);
+    expect(response.data.category.name).toEqual(randomFirstName);
+    expect(response.data.name).toEqual(randomSecondName);
    });
 
 
-test('Delete', async () => {
-    const response = await axios.delete(`${baseUrl}${petId}`);
-    expect(response.status).toBe(200);
-});
