+++ tests/api_tests_pet_store/put.spec.ts
@@ -1,6 +1,6 @@
 import {test, expect} from '@playwright/test'
 import axios from 'axios';
-import { baseUrl } from '../Global_settings/constant';
+import { baseUrl } from '../global_settings/constant';
 import {faker, simpleFaker} from '@faker-js/faker';
 
 
@@ -8,7 +8,7 @@ let randomFirstName = faker.person.firstName();
 let randomSecondName = faker.person.lastName();
 let randomId = simpleFaker.string.numeric(10);
 let updatedFirstName = faker.person.firstName();
-let updatedSecondName = faker.person.firstName();
+let updatedSecondName = faker.person.lastName();
 let petId: number;
 
 
@@ -34,8 +34,8 @@ let postData = {
 
 test.beforeAll(async () => {
     const response = await axios.post(baseUrl, postData);
-    expect(response.status).toBe(200);
     petId = response.data.id;
+    expect(response.status).toBe(200);
    });
 
 test.afterAll(async () => {
@@ -46,7 +46,6 @@ test.afterAll(async () => {
 test('PutRequest', async () => {
     const response = await axios.put(baseUrl, 
     { 
-      data: {
         "id": petId,
         "category": {
           "id": 0,
@@ -62,16 +61,11 @@ test('PutRequest', async () => {
             "name": "string"
           }
         ],
-        "status": "available"}
+        "status": "available"
     });
     expect(response.status).toBe(200);
+    expect(response.data.category.name).toEqual(updatedFirstName);
+    expect(response.data.name).toEqual(updatedSecondName);
    });
-
-   test ('Get', async () => {
-    const response = await axios.get(`${baseUrl}${petId}`);
-    expect(response.data.category.name).toEqual(randomFirstName);
-    expect(response.data.name).toEqual(randomSecondName);
-    expect(response.status).toBe(200);
-});
