+++ tests/global_settings/constant.ts
@@ -1,0 +1,1 @@
+export let baseUrl = 'https://petstore.swagger.io/v2/pet/';
